% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyvis2-package.R
\docType{package}
\name{phyvis2-package}
\alias{phyvis2}
\alias{phyvis2-package}
\title{phyvis2: Phylogenetic Analysis and Visialization}
\description{
The fucntion in this package allows you to explore different areas of phylogenetic analyses with the crucial component of visualization. This function will plot the phylogenetic tree using ggtree(). Tip labels, a tree scale, and a legend have been added. The branch lengths have been colored in coral and black.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/electrickearns/phyvis2.git}
  \item Report bugs at \url{https://github.com/electrickearns/phyvis2/issues}
}

}
\keyword{internal}
