% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.vis.R
\name{tree.vis}
\alias{tree.vis}
\title{tree.vis}
\usage{
tree.vis(x)
}
\arguments{
\item{x}{- data}
}
\description{
This function creates a simple phylogenetic tree, complete with branch length. It relies on the K80 model of distance
}
\examples{
# require (ape)
# require(ggtree)
# tree.vis(MySeq)

}
\keyword{-}
\keyword{DNA}
